demo/customers/views.py
from django.contrib.auth.models import User
from rest_framework import viewsets
from .serializers import CustomerSerializer


class CustomerViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = CustomerSerializer
